server/backend/requirements.txt
Flask==3.0.3
mysql-connector-python==8.4.0
waitress==3.0.0
